requirements.txt
streamlit>=1.24.0
agno>=0.1.0
openai>=1.0.0
python-dotenv>=0.19.0
nest-asyncio>=1.5.6
anyio>=3.7.1
google-api-python-client==2.118.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
requests==2.31.0
fastmcp==0.3.3
aiohttp==3.12.0


